# frontend/src/api/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Sofumar Portal API — Auto-generated OpenAPI specification.",
    "description" : "!!! DO NOT EDIT THIS FILE MANUALLY !!!\nThis file is auto-generated by automation pipeline. Any manual changes will be lost during the next build.",
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:50437",
    "description" : "Generated server url"
  } ],
  "security" : [ {
    "bearerAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/members/update" : {
      "put" : {
        "tags" : [ "Members" ],
        "summary" : "Update an existing member",
        "operationId" : "updateMember",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/update" : {
      "put" : {
        "tags" : [ "Expenses" ],
        "summary" : "Update an existing expense",
        "operationId" : "updateExpense",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/members/search" : {
      "post" : {
        "tags" : [ "Members" ],
        "summary" : "Search members",
        "operationId" : "searchMembers",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberSearchRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListMemberDto"
                }
              }
            }
          }
        }
      }
    },
    "/members/add" : {
      "post" : {
        "tags" : [ "Members" ],
        "summary" : "Add a new member",
        "operationId" : "addMember",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MemberDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseInteger"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/search" : {
      "post" : {
        "tags" : [ "Expenses" ],
        "summary" : "Search expenses",
        "operationId" : "searchExpenses",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseSearchRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListExpenseDto"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/add" : {
      "post" : {
        "tags" : [ "Expenses" ],
        "summary" : "Add a new expense",
        "operationId" : "addExpense",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ExpenseDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseInteger"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/register" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Register a new user",
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/auth/password-update" : {
      "post" : {
        "tags" : [ "Authentication" ],
        "summary" : "Update current user password",
        "operationId" : "updatePassword",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PasswordUpdateRequestDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/members/{memberID}/summary" : {
      "get" : {
        "tags" : [ "Members" ],
        "summary" : "Get member summary (total paid, outstanding, overdue)",
        "operationId" : "getMemberSummary",
        "parameters" : [ {
          "name" : "memberID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseMemberSummaryDto"
                }
              }
            }
          }
        }
      }
    },
    "/members/lookup" : {
      "get" : {
        "tags" : [ "Members" ],
        "summary" : "Lookup members - fuzzy search by first name, last name or ID",
        "operationId" : "lookupMembers",
        "parameters" : [ {
          "name" : "query",
          "in" : "query",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseListMemberLookupDto"
                }
              }
            }
          }
        }
      }
    },
    "/members/get/{memberID}" : {
      "get" : {
        "tags" : [ "Members" ],
        "summary" : "Get member by ID",
        "operationId" : "getMember",
        "parameters" : [ {
          "name" : "memberID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseMemberDto"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/get/{expenseID}" : {
      "get" : {
        "tags" : [ "Expenses" ],
        "summary" : "Get expense by ID",
        "operationId" : "getExpense",
        "parameters" : [ {
          "name" : "expenseID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseExpenseDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/members/delete/{memberID}" : {
      "delete" : {
        "tags" : [ "Members" ],
        "summary" : "Delete member by ID",
        "operationId" : "deleteMember",
        "parameters" : [ {
          "name" : "memberID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/expenses/delete/{expenseID}" : {
      "delete" : {
        "tags" : [ "Expenses" ],
        "summary" : "Delete expense by ID",
        "operationId" : "deleteExpense",
        "parameters" : [ {
          "name" : "expenseID",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GlobalResponseVoid"
                }
              }
            }
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
...
      "GlobalResponseVoid" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "description" : "Response payload"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
      "MemberDto" : {
        "type" : "object",
        "properties" : {
          "memberID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "joinDate" : {
            "type" : "string",
            "format" : "date"
          },
          "address1" : {
            "type" : "string"
          },
          "address2" : {
            "type" : "string"
          },
          "city" : {
            "type" : "string"
          },
          "state" : {
            "type" : "string"
          },
          "zip" : {
            "type" : "string"
          }
        }
      },
      "ExpenseDto" : {
        "type" : "object",
        "properties" : {
          "expenseID" : {
            "type" : "integer",
            "format" : "int32"
          },
          "dateOfExpense" : {
            "type" : "string",
            "format" : "date"
          },
          "category" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "amount" : {
            "type" : "number"
          }
        }
      },
      "SortOrder" : {
        "type" : "string",
        "enum" : [ "asc", "desc" ]
      },
...
      "GlobalResponseInteger" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "Response payload"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
      "MemberSearchRequestDto" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Page number (0-based)",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "description" : "Page size",
            "example" : 10
          },
          "sortField" : {
            "type" : "string",
            "description" : "Sort field name",
            "example" : "firstName"
          },
          "sortOrder" : {
            "$ref" : "#/components/schemas/SortOrder",
            "description" : "Sort direction",
            "example" : "asc"
          },
          "firstName" : {
            "type" : "string"
          },
          "lastName" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        }
      },
      "GlobalResponseListMemberDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/MemberDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
      "ExpenseSearchRequestDto" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0,
            "description" : "Page number (0-based)",
            "example" : 0
          },
          "size" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 10,
            "description" : "Page size",
            "example" : 10
          },
          "sortField" : {
            "type" : "string",
            "description" : "Sort field name",
            "example" : "firstName"
          },
          "sortOrder" : {
            "$ref" : "#/components/schemas/SortOrder",
            "description" : "Sort direction",
            "example" : "asc"
          },
          "category" : {
            "type" : "string"
          },
          "dateFrom" : {
            "type" : "string",
            "format" : "date"
          },
          "dateTo" : {
            "type" : "string",
            "format" : "date"
          }
        }
      },
      "GlobalResponseListExpenseDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/ExpenseDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
      "UserDto" : {
        "type" : "object",
        "properties" : {
          "firstName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 1
          },
          "lastName" : {
            "type" : "string",
            "maxLength" : 100,
            "minLength" : 1
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "minLength" : 1
          },
          "username" : {
            "type" : "string",
            "minLength" : 1
          },
          "password" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 5
          },
          "role" : {
            "type" : "string"
          }
        },
        "required" : [ "email", "firstName", "lastName", "password", "username" ]
      },
      "PasswordUpdateRequestDto" : {
        "type" : "object",
        "properties" : {
          "oldPassword" : {
            "type" : "string",
            "minLength" : 1
          },
          "newPassword" : {
            "type" : "string",
            "maxLength" : 50,
            "minLength" : 5
          },
          "confirmPassword" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "confirmPassword", "newPassword", "oldPassword" ]
      },
...
      "GlobalResponseMemberSummaryDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "$ref" : "#/components/schemas/MemberSummaryDto",
            "description" : "Response payload"
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
      "GlobalResponseListMemberLookupDto" : {
        "type" : "object",
        "description" : "Unified API response envelope",
        "properties" : {
          "statusCode" : {
            "type" : "integer",
            "format" : "int32",
            "description" : "HTTP status code",
            "example" : 200
          },
          "statusDesc" : {
            "type" : "string",
            "description" : "HTTP status description",
            "example" : "OK"
          },
          "globalMessages" : {
            "type" : "array",
            "description" : "Global-level messages (errors, warnings, info)",
            "items" : {
              "$ref" : "#/components/schemas/GlobalMsg"
            }
          },
          "fieldMessages" : {
            "type" : "array",
            "description" : "Field-level validation messages",
            "items" : {
              "$ref" : "#/components/schemas/FieldMsg"
            }
          },
          "responseData" : {
            "type" : "array",
            "description" : "Response payload",
            "items" : {
              "$ref" : "#/components/schemas/MemberLookupDto"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PaginationMeta",
            "description" : "Pagination metadata"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "bearerAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}